+++ lib/widgets/icon/submenu_with_icon.dart
@@ -2,11 +2,16 @@ part of '../../ipados_menu_bar.dart';
 
 /// Custom [PlatformMenu] class with a new ***icon*** attribute.
 class PlatformMenuWithIcon extends PlatformMenu {
-  IconData icon;
+  IconData? icon;
+  Widget? iconWidget;
 
   PlatformMenuWithIcon({
-    required this.icon,
+    this.icon,
+    this.iconWidget,
     required super.label,
     required super.menus,
-  });
+  }) : assert(
+         (icon == null) != (iconWidget == null),
+         'Exactly one of icon or iconWidget must be provided.',
+       );
 }